Move find to iter
