+++ iter/iter_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/BooleanCat/go-functional/internal/assert"
 	"github.com/BooleanCat/go-functional/iter"
 	"github.com/BooleanCat/go-functional/iter/ops"
+	"github.com/BooleanCat/go-functional/option"
 )
 
 func ExampleCollect() {
@@ -33,6 +34,14 @@ func ExampleToChannel() {
 	// 1
 	// 2
 	// 3
+}
+
+func ExampleFind() {
+	values := iter.Lift([]string{"foo", "bar", "baz"})
+	bar := iter.Find[string](values, func(v string) bool { return v == "bar" })
+
+	fmt.Println(bar)
+	// Output: Some(bar)
 }
 
 func TestCollect(t *testing.T) {
@@ -88,4 +97,19 @@ func TestForEachEmpty(t *testing.T) {
 	})
 
 	assert.Empty[string](t, sum)
+}
+
+func TestFind(t *testing.T) {
+	values := iter.Lift([]string{"foo", "bar", "baz"})
+	bar := iter.Find[string](values, func(v string) bool { return v == "bar" })
+
+	assert.Equal(t, bar, option.Some("bar"))
+	assert.Equal(t, values.Next().Unwrap(), "baz")
+}
+
+func TestFindEmpty(t *testing.T) {
+	values := iter.Exhausted[int]()
+	found := iter.Find[int](values, func(v int) bool { return v == 0 })
+
+	assert.True(t, found.IsNone())
 }
